add TherapySettingExtension
